Rename campaign component to Campaigns in order to notice capitalization
Git didn't want to recognize that campaign and Campaign are two
different things. So now it's called Campaigns.
